# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API приложения для тестового задания",
        "description": "API-документация для CRUD организаций, зданий, видов деятельности и аутентификации.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:80/api",
            "description": "API сервер"
        }
    ],
    "paths": {
        "/activities": {
            "get": {
                "tags": [
                    "Activities"
                ],
                "summary": "Получить все деятельности в бд",
                "operationId": "b14ed4d53a740144ff049ec7cc961f2c",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "15"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список деятельностей",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Activities"
                ],
                "summary": "Создать новую активность",
                "operationId": "17ca1fdfc4a2bdfc45cf306f694fa5e1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Строительство"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": "null"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Деятельность создана",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Превышен максимальный уровень вложенности (3)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/activities/{id}": {
            "put": {
                "tags": [
                    "Activities"
                ],
                "summary": "Обновить деятельность",
                "operationId": "9cce514da02ea827e366ff51cf7bd8ce",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Автомобили"
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": "null"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Деятельность обновлена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Превышен максимальный уровень вложенности (3)"
                    },
                    "404": {
                        "description": "Деятельность не найдена"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Activities"
                ],
                "summary": "Удалить деятельность",
                "operationId": "37ae59b21572084eada983b39be30e04",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Деятельность удалена",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/buildings": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить все здания в бд",
                "operationId": "8dbf4d765ceccf748097033980369501",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "15"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список зданий",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Добавить новое здание",
                "operationId": "91828bb17342458c9c9e1d3382ef3198",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address",
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "1-й Зачатьевский пер., д.6ст1"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "55.74039"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "37.60216"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Деятельность создана",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/buildings/{id}": {
            "put": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Обновить данные о здании",
                "operationId": "fbc9958f26c4410b19f5b059bec2e49b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address",
                                    "lat",
                                    "lng"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "1-й Зачатьевский пер., д.6ст1"
                                    },
                                    "lat": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "55.74039"
                                    },
                                    "lng": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "37.60216"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Здание обновлено",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Удалить здание",
                "operationId": "d5e7526ffccb2953f696d47de5b43656",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Здание удалено",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}